Move FluxPointsDataset to gammapy/datasets
